requirement.txt.txt
django==2.2
djangorestframework-3.12.2